+++ src/server/plugins/vatsim.ts
@@ -1,8 +1,14 @@
 import { CronJob } from 'cron';
-import type { VatsimData, VatsimDivision, VatsimEvent, VatsimSubDivision } from '~/types/data/vatsim';
+import type {
+    VatsimData,
+    VatsimDivision,
+    VatsimEvent,
+    VatsimSubDivision,
+    VatsimTransceiver,
+} from '~/types/data/vatsim';
 import { radarStorage } from '~/utils/backend/storage';
 import { getAirportsList, getATCBounds, getLocalATC, useFacilitiesIds } from '~/utils/data/vatsim';
-import { fromServerLonLat } from '~/utils/backend/vatsim';
+import { fromServerLonLat, getTransceiverData } from '~/utils/backend/vatsim';
 
 function excludeKeys<S extends {
     [K in keyof D]?: D[K] extends Array<any> ? {
@@ -30,17 +36,18 @@ function excludeKeys<S extends {
 }
 
 export default defineNitroPlugin(app => {
-    let latestFinished = 0;
-    let isInProgress = false;
+    let dataLatestFinished = 0;
+    let dataInProgress = false;
+    let transceiversInProgress = false;
 
     CronJob.from({
         cronTime: '* * * * * *',
         start: true,
         runOnInit: true,
         onTick: async () => {
-            if (!radarStorage.vatspy.data || isInProgress || Date.now() - latestFinished < 1000) return;
+            if (!radarStorage.vatspy.data || dataInProgress || Date.now() - dataLatestFinished < 1000) return;
             try {
-                isInProgress = true;
+                dataInProgress = true;
                 const data = await $fetch<VatsimData>('https://data.vatsim.net/v3/vatsim-data.json', {
                     parseResponse(responseText) {
                         return JSON.parse(responseText);
@@ -147,8 +154,8 @@ export default defineNitroPlugin(app => {
                 console.error(e);
             }
             finally {
-                isInProgress = false;
-                latestFinished = Date.now();
+                dataInProgress = false;
+                dataLatestFinished = Date.now();
             }
         },
     });
@@ -182,6 +189,30 @@ export default defineNitroPlugin(app => {
         runOnInit: true,
         onTick: async () => {
             radarStorage.vatsim.events = (await $fetch<{ data: VatsimEvent[] }>('https://my.vatsim.net/api/v2/events/latest')).data;
+        },
+    });
+
+    CronJob.from({
+        cronTime: '* * * * * *',
+        start: true,
+        runOnInit: true,
+        onTick: async () => {
+            if (!radarStorage.vatspy.data || transceiversInProgress) return;
+            try {
+                transceiversInProgress = true;
+                radarStorage.vatsim.transceivers = await $fetch<VatsimTransceiver[]>('https://data.vatsim.net/v3/transceivers-data.json', {
+                    parseResponse(responseText) {
+                        return JSON.parse(responseText);
+                    },
+                    timeout: 1000 * 30,
+                });
+            }
+            catch (e) {
+                console.error(e);
+            }
+            finally {
+                transceiversInProgress = false;
+            }
         },
     });
 });